+++ src/Services/TwitchApiService.php
@@ -21,8 +21,15 @@ use Zarlach\TwitchApi\API\Videos;
 class TwitchApiService extends Api
 {
     /**
-     * Authentication.
-     */
+    * Authentication.
+    */
+    public function getAuthenticationUrl($state = null, $forceVerify = false)
+    {
+        $authentication = new Authentication();
+
+        return $authentication->getAuthenticationUrl($state, $forceVerify);
+    }
+
     public function getAccessToken($code, $state = null)
     {
         $authentication = new Authentication();
@@ -30,11 +37,11 @@ class TwitchApiService extends Api
         return $authentication->getAccessToken($code, $state);
     }
 
-    public function getAuthenticationUrl($state = null, $forceVerify = false)
+    public function getAccessToken($code, $state = null)
     {
         $authentication = new Authentication();
 
-        return $authentication->getAuthenticationUrl($state, $forceVerify);
+        return $authentication->getAccessObject($code, $state);
     }
 
     /**